# [narrowing] Make subsequent .with clause inherit narrowing