--- a/tests/type-error.test.ts
+++ b/tests/type-error.test.ts
@@ -20,12 +20,14 @@ describe('type errors', () => {
     match<Country>('France')
       // @ts-expect-error: 'Spai' instead of 'Spain'
       .with('France', 'Germany', 'Spai', (x) => 'Europe')
+      // @ts-expect-error
       .exhaustive();
 
     match<Country>('Germany')
       .with('Germany', 'Spain', () => 'Europe')
       // @ts-expect-error: 'US' instead of 'USA'
       .with('US', (x) => 'America')
+      // @ts-expect-error
       .exhaustive();
   });
 
@@ -37,6 +39,7 @@ describe('type errors', () => {
         type t = Expect<Equal<typeof x, 'France' | 'USA'>>;
         return 'America';
       })
+      // @ts-expect-error
       .exhaustive();
   });
 
@@ -46,12 +49,16 @@ describe('type errors', () => {
         .with({ kind: 'some', value: { x: 2 } }, () => '2')
         // @ts-expect-error, value.x should be a number
         .with({ value: { x: '' } }, () => '2')
+        .with({ kind: 'none' }, () => '')
+        .with({ kind: 'some' }, () => '')
         .exhaustive();
 
     const f2 = (input: Option<number>) =>
       match(input)
         // @ts-expect-error: value is a number
         .with({ kind: 'some', value: 'string' }, () => '')
+        .with({ kind: 'none' }, () => '')
+        .with({ kind: 'some' }, () => '')
         .exhaustive();
   });